RF: Move no_annex() out of plugins
Towards a fix for #4519
